+++ lib/mplcursors/__init__.py
@@ -10,9 +10,9 @@ except (ImportError, LookupError):
         pass
 
 
-from ._mplcursors import Cursor, cursor
+from ._mplcursors import Cursor, HoverMode, cursor
 from ._pick_info import Selection, compute_pick, get_ann_text, make_highlight
 
 
-__all__ = ["Cursor", "cursor", "Selection",
+__all__ = ["Cursor", "HoverMode", "cursor", "Selection",
            "compute_pick", "get_ann_text", "make_highlight"]